refac(91026): Extrai task gerar_notificacao_encerramento_conta_bancaria_async
